working state | ep 16

+++ Leonida, Althani Miguel/Week_4/02_mern-app1/frontend/src/components/WorkoutDetails.js
@@ -1,23 +1,32 @@
 
 import { useWorkoutsContext } from '../hooks/useWorkoutsContext';
+import { useAuthContext } from '../hooks/useAuthContext'
 
 //date-fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 
-const WorkoutDetails = ({workout}) => {
+const WorkoutDetails = ({ workout }) => {
 
-    const {dispatch} = useWorkoutsContext();
+    const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext()
 
     const handleClick = async () => {
-        const response = await fetch ('/api/workouts/' + workout._id, {
+
+        if (!user) {
+            return
+        }
+
+
+        const response = await fetch('/api/workouts/' + workout._id, {
             method: 'DELETE',
+            headers: { 'Authorization': `Bearer ${user.token}` }
         });
 
         const json = await response.json();
 
         if (response.ok) {
-            dispatch({type: 'DELETE_WORKOUT', payload: json})
+            dispatch({ type: 'DELETE_WORKOUT', payload: json })
         }
     }
     return (
@@ -28,20 +37,20 @@ const WorkoutDetails = ({workout}) => {
 
             <p>
                 <strong>
-                    Load (kg): 
+                    Load (kg):
                 </strong>
                 {workout.load}
             </p>
 
             <p>
                 <strong>
-                    Reps: 
+                    Reps:
                 </strong>
                 {workout.reps}
             </p>
 
             <p>
-                {formatDistanceToNow(new Date(workout.createdAt), {addSuffix: true})}
+                {formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}
             </p>
 
             <span className="material-symbols-outlined" onClick={handleClick}>
